requirements.txt
streamlit
python-docx
transformers
torch
huggingface_hub
